requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
defusedxml==0.7.1
Django==4.2.7
django-bootstrap-icons==0.9.0
django-browser-reload==1.18.0
django-crispy-forms==2.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.2.0
mysqlclient==2.2.7
packaging==24.2
phonenumbers==9.0.1
pillow==11.1.0
propcache==0.3.0
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.1
requests==2.31.0
sqlparse==0.5.3
stripe==7.6.0
twilio==8.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
yarl==1.18.3
